написал код для 25 задачи

+++ Task23/Program.cs
@@ -1,12 +1,12 @@
 ﻿Console.Write("Введите натуральное число: ");
-int number = Convert.ToInt32(Console.ReadLine());
+long number = Convert.ToInt64(Console.ReadLine());
 if (number < 1) Console.WriteLine("Некорректный ввод");
 else TableCubes(number);
 
 
-void TableCubes(int num)
+void TableCubes(long num)
 {
-    int count = 1;
+    long count = 1;
     while (count <= num)
     {
         Console.WriteLine($"{count, 5}   {count * count * count, 5}");
